Backend/requirements.txt
fastapi[all]
uvicorn==0.38.0
bcrypt==5.0.0
PyJWT==2.10.1
python-multipart==0.0.20
google-generativeai==0.8.5
sqlmodel==0.0.27
